requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-learn>=1.0.0
pyyaml>=5.4.0

# ECG/EEG processing
neurokit2>=0.2.0

# Visualization
seaborn>=0.11.0

# Optional: for better performance
tqdm>=4.62.0